successfully added diff_in_curv metrics to planning_data_by_point. Will work on incorporating it into dash

+++ multiff_code/methods/planning_analysis/show_planning/cur_vs_nxt_ff/find_cvn_utils.py
@@ -542,7 +542,7 @@ def find_ff_info_n_cm_ago(ff_df, monkey_information, ff_real_position_sorted, n_
 
 def find_ff_info(all_ff_index, all_point_index, monkey_information, ff_real_position_sorted):
     ff_df = pd.DataFrame(
-        {'ff_index': all_ff_index, 'point_index': all_point_index})
+        {'ff_index': all_ff_index.astype(int), 'point_index': all_point_index.astype(int)})
 
     ff_df[['ff_x', 'ff_y']] = ff_real_position_sorted[ff_df['ff_index'].values]
     ff_df[['monkey_x', 'monkey_y', 'monkey_angle']] = monkey_information.loc[ff_df['point_index'], [
